make: use vendored goimports to format generated APIs
[ upstream commit 8606b0cd1e9279cf2c5bee3bbe2e9eeb8c05b745 ]

Commit 563787ebb318 ("makefile: use versioned go container when
formatting after api generate.") changed the API generation make targets
to use a container running a pinned version of goimports. This can be
achieved more easily by running the vendored version of goimports using
the native Go toolchain. This also makes sure that the version is
automatically kept up-to-date in go.mod using renovatebot instead of
having to manually bump GOIMPORTS_VERSION.

For #24806

Signed-off-by: Tobias Klauser <tobias@cilium.io>