gha: extend conformance clustermesh to also cover external kvstores
[ upstream commit 403b3a265085ea3e08c72b59ee2d5dd2836fe4a4 ]

Let's extend the conformance clustermesh workflow to also cover the
external kvstores configuration in addition to plain clustermesh and
kvstoremesh. To avoid increasing the number of matrix entries, let's
convert two of the already existing ones over to this mode.

We leverage the newly introduced kvstore action to setup the etcd
containers and retrieve the appropriate parameters. Cluster Mesh
configurations are directly specified at installation time, as
'cilium clustermesh connect' does not support this scenario.

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>