gha: prevent circular dependency in clustermesh-upgrade workflow
[ upstream commit 00ed827d9c6f7962df34f3a500468a1b42c7f2f4 ]

The simultaneous restart of the clustermesh-apiserver pods in both
clusters after rolling out all agents can lead to a circular dependency
when Cilium is configured in tunneling mode and KPR=true [1]. For the
moment, let's avoid to trigger this scenario in CI, as unlikely to
happen in real environments. We never hit this issue before because
we only had one worker node, which is targeted by the NodePort, and
apparently the clustermesh-apiserver was always scheduled there.

[1]: https://github.com/cilium/cilium/issues/30156

Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>