helm: Support extending hubble-relay helm templates as a downstream packager
Extending hubble-relay with a sidecar is currently possible by supplying
additional values, but this only works for cases where the users know
what values to use to extend it. We want to repackage the cilium chart
but add an additional sidecar proxy in front of hubble-relay with
minimal modifications to the upstream chart manifests.

Currently, adding new default sidecar containers into
hubble-relay requires directly adding them to the chart or patching it
after it's been rendered by helm.

Introducing new sidecars upstream feels inappropriate since the ones we
want to add are not generally applicable to all users, and requiring users
to patch the manifests after helm (eg: using kustomize) is cumbersome
and prone to error.

In order to support adding new default sidecars to hubble-relay, I have
introduced new template blocks in a new _extensions.tpl to allow
downstream packagers to more easily modify the default containers that
hubble-relay runs with.

To fully support running a sidecar proxy in-front of hubble-relay, a few
things is needed:
- new sidecar containers that can be enabled conditionally
- volumes that the sidecar uses
- changing the listen address/listen port for hubble-relay so the proxy
  can handle the requests before handing them off to hubble-relay
- changing the TLS configuration of hubble-relay.

To support each of these modfications, a new template block ("partial")
has been added allowing downstream packagers to simplify modify
_extensions.tpl in their fork with the required logic to enable the new
sidecar.

Signed-off-by: Chance Zibolski <chance.zibolski@gmail.com>