l2announcer: Retry getting lease after losing it
[ upstream commit 00b1e5e6031630cf5a8ab3da698505504b5da62d ]

Once a service gets selected we start leader election. However, if
we lose the lease for some reason, we don't retry getting it until
the service is deselected and reselect, recreated or the agent
restarts. This commit surrounds the lease leader election logic with
a loop that ends when the context is cancelled.

Signed-off-by: Dylan Reimerink <dylan.reimerink@isovalent.com>
Signed-off-by: Marco Iorio <marco.iorio@isovalent.com>